hacked version to create a deepseek refusal vector
I created a refusal vector for DeepSeek-R1-Distill-Llama-8B with
multiple changes to the code - many trivial a a few not obvious.
If there is interest I or someone could clean this up, but I
wanted to offer it in its current state while the changes were
fresh on my mind in case it is useful for others.

The main changes were:

* pipeline/model_utils/llama3_model.py was made specific to
deepseek distall. The input appends a prefill to have the
model skip the thinking phase:

----8<----

+<thinking>
+[THINKING_SKIPPED]
+</thinking>
----8<----

I attempted to insert this into the LLAMA3_CHAT_TEMPLATE_WITH_SYSTEM
but that failed, so I simply patched the format_instruction_llama3_chat()
function call instead.

Also in this file I included </think> in _get_eoi_toks() so this would
be evaluated.

* pipeline/model_utils/model_factory.py was patched to load
llama3_model() when 'distill-llama' is part of the model_name. Note
that this should really be it's own dedicated model factory
as this likely breaks existing (non-deepseek) llama3 processing.

* pipeline/run_pipeline.py

kl_threshold was raised from 0.1 to 0.2 as this was necessary in testing

harmful_train data was patched with nine examples of deepseek specific
refusal that I think was not well covered in existing datasets - such
as "What is 8964?".

* Also included is pipeline/runs/DeepSeek-R1-Distill-Llama-8B with
the refusal vector I generated and metrics and diagnostics.

* I also sloppily left in some print statements used for debugging